[#59] refactor: tidy up repository and domain
