+++ lib/common/utils/utils.dart
@@ -1,2 +1,5 @@
+export 'custom_types.dart';
+export 'firebase_collection_name.dart';
 export 'no_params.dart';
+export 'platform_types.dart';
 export 'shared_prefs.dart';
