+++ lib/features/live_writing/data/datasources/confirm_post_datasource_provider.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wehavit/features/live_writing/data/datasources/datasources.dart';
+
+final confirmPostDatasourceProvider = Provider<ConfirmPostDatasource>((ref) {
+  return ConfirmPostRemoteDatasourceImpl();
+});
