+++ lib/features/live_writing/domain/usecases/confirm_post_usecase_provider.dart
@@ -1,0 +1,14 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:wehavit/features/live_writing/domain/domain.dart';
+
+final getAllPostUseCaseProvider = Provider<GetAllPostUseCase>(
+  (ref) => GetAllPostUseCase(
+    ref.watch(confirmPostRepositoryProvider),
+  ),
+);
+
+final createPostUseCaseProvider = Provider<CreatePostUseCase>(
+  (ref) => CreatePostUseCase(
+    ref.watch(confirmPostRepositoryProvider),
+  ),
+);
